+++ frontend/src/components/modals/index.js
@@ -11,3 +11,4 @@ export { default as ConfirmModal } from './ConfirmModal.vue';
 export { default as StationFormModal } from './StationFormModal.vue';
 export { default as PlatformFormModal } from './PlatformFormModal.vue';
 export { default as InstrumentFormModal } from './InstrumentFormModal.vue';
+export { default as ExportModal } from './ExportModal.vue';